Backends interface: move the backends to a special directory.
