+++ test/test_cuda_functions.py
@@ -1,9 +1,12 @@
 import numpy as np
+
+import kernel_tuner
 from .context import skip_if_no_cuda
 from .test_runners import env
 
 import pytest
-from kernel_tuner import nvcuda, tune_kernel
+from kernel_tuner import tune_kernel
+from kernel_tuner.backends import nvcuda
 from kernel_tuner.core import KernelSource, KernelInstance
 
 try:
